Add script that monitors lighttpd connection count
Lately there's been several instances where the tool
became unavailable because of a large number of
connections stuck in CLOSE_WAIT state, see
https://phabricator.wikimedia.org/T104799

The status page does not become available, so we
can check that to get the number of active fcgi
connections. If the number exceeds 150, we
restart the server and write an entry to a log.

The script is supposed to be run every minute.
The connection count change over time is logged,
so we can see if the connection count increases
suddenly or more gradually over time.