Fix the 1px issue with magic border locator (#59)
With the jcrop_api.setSelect([x1,y1,x2,y2]) method, the values x2 and y2 are not the last pixels inside the selection, but the first pixels outside the selection. As a result, jcrop_api.setSelect([0,0,1,1]) is a 1x1 selection, not a 2x2. So we need to add 1 px for both.