Proof of concept: use structured data regions
Using Structured Data on Commons, users can define which subjects are
depicted in an image, and which region of the image they reside in. (One
way to define these regions is the Wikidata Image Positions tool [1],
which despite its name also supports Commons.) We can use these regions
as a basis for cropping – for example, if an editor has already marked
where individual people are located in a group photo, the same data can
be used to crop out images of those individuals.

[1]: https://www.wikidata.org/wiki/User:Lucas_Werkmeister/Wikidata_Image_Positions