Fix and use the JpegFile class for lossless cropping
The refactoring done in 382d7cc1 left the JpegFile class in an un-
usuable state.

Using the classname alone to do a static function call on the crop
method in the Image class could not work when the path of jpegtran
is configured on the object.

That's the reason why lossless jpeg manipulation did not work.

The FileRepository actually does a good job creating objects that
fit the file type so they could be used directly in the Image class
and so the static call there can just be a call on the specific
instance.